+++ client-core/src/test/java/de/njsm/stocks/client/business/event/EventInteractorImplTest.java
@@ -83,8 +83,8 @@ public class EventInteractorImplTest {
         when(repository.getFoodFeed(any())).thenReturn(Single.just(emptyList()));
         when(repository.getFoodItemFeed(any())).thenReturn(Single.just(emptyList()));
         when(repository.getEanNumberFeed(any())).thenReturn(Single.just(emptyList()));
-        when(repository.getPreviousDayContainingEvents(any())).thenReturn(Maybe.empty());
-        when(repository.getNextDayContainingEvents(any())).thenReturn(Maybe.empty());
+        when(repository.getPreviousDayContainingEvents(any(), any())).thenReturn(Maybe.empty());
+        when(repository.getNextDayContainingEvents(any(), any())).thenReturn(Maybe.empty());
     }
 
     @AfterEach