+++ client-database-android/src/androidTest/java/de/njsm/stocks/client/database/SqliteVersionTest.java
@@ -32,6 +32,6 @@ public class SqliteVersionTest extends DbTestCase {
 
     @Test
     public void verifyRecentSqliteVersion() {
-        assertEquals("3.36.0", stocksDatabase.metadataDao().getSqliteVersion());
+        assertEquals("3.42.0", stocksDatabase.metadataDao().getSqliteVersion());
     }
 }